Backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
google-generativeai
python-dotenv
